update variable names and comments for readability
