+++ src/utils/manageTerminals.ts
@@ -1,10 +1,10 @@
 import * as vscode from "vscode";
 
-import runCommands from "../utils/runCommands";
+import runAction from "./runAction";
 import getConfigFile from "../utils/getConfigFile";
 
 /**
- * This function retrieves the config file and runs the commands associated with the action.
+ * Retrieves the config file and runs the selected action.
  * If the action is not found in the configuration file, it shows an error message.
  *
  * @param action - The action to perform. This should be a key in the terminal.config.json file.
@@ -17,7 +17,7 @@ export default function manageTerminals(action: string) {
   }
 
   if (configFile[action]) {
-    runCommands(configFile[action]);
+    runAction(configFile[action]);
   } else {
     vscode.window.showErrorMessage(
       `Action not found in the configuration file: ${action}`
